+++ src/HCM.Application/Auth/Commands/Logout/LogoutCommand.cs
@@ -1,5 +1,0 @@
-﻿using MediatR;
-
-namespace HCM.Application.Auth.Commands.Logout;
-
-public record LogoutCommand(string RefreshToken) : IRequest;